+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentinvokeop.py
@@ -59,6 +59,8 @@ Provider = Literal[
     "poolside",
     "tencent",
     "nebius",
+    "fireworks",
+    "baseten",
     "reson8",
     "slack",
     "orq",