+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -507,9 +507,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-27T09:51:24.480Z"
+    created: Optional[str] = "2026-07-28T04:12:16.755Z"
 
-    updated: Optional[str] = "2026-07-27T09:51:24.480Z"
+    updated: Optional[str] = "2026-07-28T04:12:16.755Z"
 
     updated_by_id: OptionalNullable[str] = UNSET