+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYHFSEQ3C33SA1ZT94MSY4F7"
+        "01KYKESHR2PM9PR2MYAS523RAA"
     )
     r"""The unique identifier of the data source"""