+++ packages/orq-rc/RELEASES.md
@@ -9759,3 +9759,13 @@ Based on:
 - [python v4.13.0-rc.7] packages/orq-rc
 ### Releases
 - [PyPI v4.13.0-rc.7] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.7 - packages/orq-rc
+
+## 2026-07-28 04:14:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.8] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.8] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.8 - packages/orq-rc